Create Two Dimensional Arrays
A bit trickier, this one. Write an application that creates a two-dimensional array of Strings, with two rows and three columns. Print the value in the second row and third column.

Hint: you can create an array of strings just like you created an array of floating point values above, but with text in quotes (e.g. "I'm a string") instead of floating-point numbers.

To define a 2D array, remember that the symbol [] means a 1D array (compare the answer to the exercise above), while [][] means a 2D array.

Also, remember that when you set the values of the array, each of the values is itself an array!

Finally, if this has you smashing plates and climbing the wall, check the answer, then hide it again and see if you can remember how it works.