Create Looping Through 2D Arrays
Create an application that uses two nested for loops to loop through the 2D array defined above and print the values.

Hint: Here's the outer loop:

for(int row = 0; row < 2; row++) {
}

The inner loop should use a loop variable called "col" (or whatever you like) and should loop through the columns.

Hard code the number of columns and rows, as above.

Once again, this is one of the trickiest beginner's tasks in Java. If you can't work it out, check the answer, then hide it again and try to write it from memory.