Create Improved Use of Scanner
The above program crashes if a user enters something other than a number. The problem is that we use the nextInt() method of Scanner, assuming blindly that we will get an integer.

Modify the program so that no input can crash it.

Hint: you will need the hasNextInt() method of Scanner. You will also need the nextLine() method so that you can get a line from the user if they don't enter an integer.

You'll need to use an if...else statement that detects if the user enters an int, and simply gets a line (rather than an int) if he doesn't. Don't be afraid to consult the API docs for the Scanner class; type "java api scanner" into Google (without the quotes).